requirements.txt
aws-xray-sdk==2.14.0
blinker==1.9.0
boto3==1.37.7
botocore==1.37.7
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.11.4
six==1.17.0
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
